Import templates on startup #3
The import folder is /project/import/template.
If this folder is empty, the process should be skipped.
The command name for import is project:overwrite-with-import.
The command flag --force must be present.
The file format must be json.
A malformed file must be skipped.
Clear the templates from the storage before the import is started.
Environment variable APPLICATION_OVERWRITE_WITH_IMPORT_ON_STARTUP
must be set to true in order to import when the entrypoint
of the container is running.